requirements.txt
pip==23.1.2
certifi==2021.5.30
Flask_Cors==3.0.10
Flask_PyMongo==2.3.0
matplotlib==3.5.1
pandas==1.3.5
plotly==5.15.0
prophet==1.1.0
pymongo==4.1.1
python-dateutil==2.8.2
scikit-learn==1.0.2
seaborn==0.11.2
gunicorn==20.1.0
holidays==0.22
dnspython==2.3.0